Refactor env to store objects *only* in the hashmap
Key design notes:

A state variable is used to key the "last", or final object in the current evaluation for the environment to the store. This is done to avoid duplicating this object both inside and outside of the store, and also because I don't think it's possible to store a reference to a another struct field within the same struct in safe rust.

This leaves us with having to maintain the invariant that return_state should always be a valid indicator of objects in the store, hence the use of panics in the code here.

Rules:
- Types used by this object should not be exposed to consumers even in the same module
- Methods should preserve immutability