Refactor bytes and instructions
Was confused over the concepts of bytecode and instructions. In this
project, we want an Instruction to be a well formed struct containing
opcodes and operands separately.

Bytes will just be a simple wrapper over a Vec<u8> that adds additional
helpers on top of it. We want to eventually turn Bytes into a iterator
for "incremental" compilation.